Refactor counter implementation into EventCounterPipeline.
