+++ src/Microsoft.Diagnostics.Monitoring.EventPipe/Counters/EventCounterPipeline.cs
@@ -3,10 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Diagnostics.NETCore.Client;
+using Microsoft.Diagnostics.Tracing;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -14,14 +13,14 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
 {
     internal class EventCounterPipeline : EventSourcePipeline<EventPipeCounterPipelineSettings>
     {
-        private readonly IEnumerable<ICountersLogger> _metricsLogger;
+        private readonly IEnumerable<ICountersLogger> _loggers;
         private readonly CounterFilter _filter;
 
         public EventCounterPipeline(DiagnosticsClient client,
             EventPipeCounterPipelineSettings settings,
-            IEnumerable<ICountersLogger> metricsLogger) : base(client, settings)
+            IEnumerable<ICountersLogger> loggers) : base(client, settings)
         {
-            _metricsLogger = metricsLogger ?? throw new ArgumentNullException(nameof(metricsLogger));
+            _loggers = loggers ?? throw new ArgumentNullException(nameof(loggers));
 
             if (settings.CounterGroups.Length > 0)
             {
@@ -37,10 +36,115 @@ namespace Microsoft.Diagnostics.Monitoring.EventPipe
             }
         }
 
-        internal override DiagnosticsEventPipeProcessor CreateProcessor()
+        internal override MonitoringSourceConfiguration CreateConfiguration()
         {
-            return new DiagnosticsEventPipeProcessor(PipeMode.Metrics, metricLoggers: _metricsLogger, metricIntervalSeconds: (int)Settings.RefreshInterval.TotalSeconds,
-                metricFilter: _filter);
+            return new MetricSourceConfiguration(CounterIntervalSeconds, _filter.GetProviders());
         }
+
+        protected override Task OnAfterEventProcessing(CancellationToken token)
+        {
+            ExecuteCounterLoggerAction((metricLogger) => metricLogger.PipelineStopped());
+
+            return Task.CompletedTask;
+        }
+
+        protected override Task OnBeforeEventProcessing(CancellationToken token)
+        {
+            ExecuteCounterLoggerAction((metricLogger) => metricLogger.PipelineStarted());
+
+            return Task.CompletedTask;
+        }
+
+        protected override void OnEventSourceAvailable(EventPipeEventSource eventSource)
+        {
+            eventSource.Dynamic.All += traceEvent =>
+            {
+                try
+                {
+                    // Metrics
+                    if (traceEvent.EventName.Equals("EventCounters"))
+                    {
+                        IDictionary<string, object> payloadVal = (IDictionary<string, object>)(traceEvent.PayloadValue(0));
+                        IDictionary<string, object> payloadFields = (IDictionary<string, object>)(payloadVal["Payload"]);
+
+                        //Make sure we are part of the requested series. If multiple clients request metrics, all of them get the metrics.
+                        string series = payloadFields["Series"].ToString();
+                        if (GetInterval(series) != CounterIntervalSeconds * 1000)
+                        {
+                            return;
+                        }
+
+                        string counterName = payloadFields["Name"].ToString();
+                        if (!_filter.IsIncluded(traceEvent.ProviderName, counterName))
+                        {
+                            return;
+                        }
+
+                        float intervalSec = (float)payloadFields["IntervalSec"];
+                        string displayName = payloadFields["DisplayName"].ToString();
+                        string displayUnits = payloadFields["DisplayUnits"].ToString();
+                        double value = 0;
+                        CounterType counterType = CounterType.Metric;
+
+                        if (payloadFields["CounterType"].Equals("Mean"))
+                        {
+                            value = (double)payloadFields["Mean"];
+                        }
+                        else if (payloadFields["CounterType"].Equals("Sum"))
+                        {
+                            counterType = CounterType.Rate;
+                            value = (double)payloadFields["Increment"];
+                            if (string.IsNullOrEmpty(displayUnits))
+                            {
+                                displayUnits = "count";
+                            }
+                            //TODO Should we make these /sec like the dotnet-counters tool?
+                        }
+
+                        // Note that dimensional data such as pod and namespace are automatically added in prometheus and azure monitor scenarios.
+                        // We no longer added it here.
+                        var counterPayload = new CounterPayload(traceEvent.TimeStamp,
+                            traceEvent.ProviderName,
+                            counterName, displayName,
+                            displayUnits,
+                            value,
+                            counterType,
+                            intervalSec);
+
+                        ExecuteCounterLoggerAction((metricLogger) => metricLogger.Log(counterPayload));
+                    }
+                }
+                catch (Exception)
+                {
+                }
+            };
+        }
+
+        private static int GetInterval(string series)
+        {
+            const string comparison = "Interval=";
+            int interval = 0;
+            if (series.StartsWith(comparison, StringComparison.OrdinalIgnoreCase))
+            {
+                int.TryParse(series.Substring(comparison.Length), out interval);
+            }
+            return interval;
+        }
+
+        private void ExecuteCounterLoggerAction(Action<ICountersLogger> action)
+        {
+            foreach (ICountersLogger logger in _loggers)
+            {
+                try
+                {
+                    action(logger);
+                }
+                catch (ObjectDisposedException)
+                {
+                }
+            }
+        }
+
+        private int CounterIntervalSeconds => (int)Settings.RefreshInterval.TotalSeconds;
     }
 }
